Code cleaned up with eclipse, some check style things
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@10823 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: 77bb522541c75346ce36e9e6b95c81eeefa6547b

+++ core/base/src/de/uniol/inf/is/odysseus/core/server/monitoring/physicaloperator/Datarate.java
@@ -62,7 +62,7 @@ public class Datarate extends AbstractPeriodicalMonitoringData<Double>
 //		synchronized (this) {
 			long currentTime = System.nanoTime();
 
-			this.value = (double) writeCount / ( (currentTime - lastTimestamp) / 1000000.0);
+			this.value = writeCount / ( (currentTime - lastTimestamp) / 1000000.0);
 			notifySubscribers(value);
 			lastTimestamp = currentTime;
 			this.writeCount = 0;
